requirements.txt
asgiref==3.8.1
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
Django==5.2.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mutagen==1.47.0
packaging==25.0
pillow==11.2.1
requests==2.32.4
sqlparse==0.5.3
urllib3==2.5.0
Werkzeug==3.1.3
yt-dlp==2025.6.9
gunicorn
cutlet
unidic-lite
ytmusicapi